# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FileApi",
    "description": "File Api V1",
    "license": {
      "name": "",
      "identifier": "MIT"
    },
    "version": "0.8"
  },
  "paths": {
    "/api/v1/template": {
      "post": {
        "tags": [
          "template::routes"
        ],
        "operationId": "upsert",
        "parameters": [
          {
            "name": "_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "title",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "description",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "templateType",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/TemplateType"
            }
          },
          {
            "name": "templateContext",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Context"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "files"
                ],
                "properties": {
                  "files": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "contentMediaType": "application/octet-stream"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Upsert a template",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateV2"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/template/find-by-context": {
      "get": {
        "tags": [
          "template::routes"
        ],
        "operationId": "find_by_context",
        "parameters": [
          {
            "name": "context",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Context"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Find templates by context",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TemplateV2"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/template/find-by-ids": {
      "get": {
        "tags": [
          "template::routes"
        ],
        "operationId": "find_by_ids",
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Find By ids",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TemplateV2"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/template/find-by-type": {
      "get": {
        "tags": [
          "template::routes"
        ],
        "operationId": "find_by_type",
        "parameters": [
          {
            "name": "template_type",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/TemplateType"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Find templates by type",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TemplateV2"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/template/find-one/{templ_id}": {
      "get": {
        "tags": [
          "template::routes"
        ],
        "operationId": "find_one",
        "parameters": [
          {
            "name": "templ_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Find a template by id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateV2"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/template/render": {
      "post": {
        "tags": [
          "template::routes"
        ],
        "operationId": "render",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Render template",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string",
                  "format": "binary",
                  "contentMediaType": "*/*"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/template/{id}": {
      "delete": {
        "tags": [
          "template::routes"
        ],
        "operationId": "delete_templ_by_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Delete a template by id"
          }
        }
      }
    },
    "/api/v1/upload": {
      "post": {
        "tags": [
          "upload::routes"
        ],
        "operationId": "upload",
        "parameters": [
          {
            "name": "correlation_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "is_public",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "boolean",
                "null"
              ]
            }
          },
          {
            "name": "without_thumbnail",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "boolean",
                "null"
              ]
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "files"
                ],
                "properties": {
                  "files": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "contentMediaType": "application/octet-stream"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Upload a file",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FileUploadV2"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/upload/download": {
      "get": {
        "tags": [
          "upload::routes"
        ],
        "operationId": "download",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Download file",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string",
                  "format": "binary",
                  "contentMediaType": "*/*"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/upload/download-bulk": {
      "post": {
        "tags": [
          "upload::routes"
        ],
        "operationId": "download_bulk",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DownloadBulkRequestUriParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Download multiple files as a zip",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string",
                  "format": "binary",
                  "contentMediaType": "*/*"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/upload/find-all": {
      "get": {
        "tags": [
          "upload::routes"
        ],
        "operationId": "find_all_uploads",
        "parameters": [
          {
            "name": "correlationId",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Find all upload",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FileUploadV2"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/upload/metadata": {
      "get": {
        "tags": [
          "upload::routes"
        ],
        "operationId": "metadata",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get upload metadata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileUploadV2"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/upload/{id}": {
      "delete": {
        "tags": [
          "upload::routes"
        ],
        "operationId": "delete_by_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Delete a file by id"
          }
        }
      }
    },
    "/api/v1/upload/{id}/make-thumb": {
      "post": {
        "tags": [
          "upload::routes"
        ],
        "operationId": "make_thumb",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Make thumbnail",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "null"
                    },
                    {
                      "$ref": "#/components/schemas/FileUploadV2"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/upload/{id}/update": {
      "post": {
        "tags": [
          "upload::routes"
        ],
        "operationId": "upload_update",
        "parameters": [
          {
            "name": "correlation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "is_public",
            "in": "path",
            "required": true,
            "schema": {
              "type": [
                "boolean",
                "null"
              ]
            }
          },
          {
            "name": "without_thumbnail",
            "in": "path",
            "required": true,
            "schema": {
              "type": [
                "boolean",
                "null"
              ]
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "files"
                ],
                "properties": {
                  "files": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "contentMediaType": "application/octet-stream"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Upload a file",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileUploadV2"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Context": {
        "type": "string",
        "enum": [
          "INVOICE",
          "PEPPOL",
          "C_V"
        ]
      },
      "DownloadBulkRequestUriParams": {
        "type": "object",
        "required": [
          "ids"
        ],
        "properties": {
          "ids": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "FileUploadV2": {
        "type": "object",
        "required": [
          "_id",
          "creationDate",
          "originalFilename",
          "size",
          "publicResource"
        ],
        "properties": {
          "_id": {
            "type": "string"
          },
          "bookmarked": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "bookmarkedDate": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "contentType": {
            "type": [
              "string",
              "null"
            ]
          },
          "correlationId": {
            "type": [
              "string",
              "null"
            ]
          },
          "creationDate": {
            "type": "string",
            "format": "date-time"
          },
          "extension": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "originalFilename": {
            "type": "string"
          },
          "publicResource": {
            "type": "boolean"
          },
          "size": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "thumb": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "thumbnailId": {
            "type": [
              "string",
              "null"
            ]
          },
          "updatedDate": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        }
      },
...
      "RenderRequest": {
        "type": "object",
        "required": [
          "templateId",
          "context",
          "fileName",
          "templateContext"
        ],
        "properties": {
          "context": {},
          "fileName": {
            "type": "string"
          },
          "templateContext": {
            "$ref": "#/components/schemas/Context"
          },
          "templateId": {
            "type": "string"
          }
        }
      },
      "TemplateType": {
        "type": "string",
        "enum": [
          "HTML",
          "XML"
        ]
      },
...
      "TemplateV2": {
        "type": "object",
        "required": [
          "_id",
          "creationDate",
          "fileId",
          "templateType",
          "templateContext",
          "title"
        ],
        "properties": {
          "_id": {
            "type": "string"
          },
          "creationDate": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "fileId": {
            "type": "string"
          },
          "templateContext": {
            "$ref": "#/components/schemas/Context"
          },
          "templateType": {
            "$ref": "#/components/schemas/TemplateType"
          },
          "title": {
            "type": "string"
          },
          "updatedDate": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        }
      },
...
    }
  }
}